Changes made to support WP7 App DataHub
Included the Dropbox concept of 'hash' for GetMetaData.
Fixed the ability to List, Upload, and Delete files at the Dropbox root.
Updated the MetaData class so the most important properties for sync could be
Serialized (to IsolatedStorage)
Added UTCDateModified for date comparison.
Added the undocumented Revision property for direct numeric comparison.