Added chunked upload auto retry on chunk failure
A default retry limit of 100 is used for now. It is per upload and not
per chunk. The retry count was added to the progress information so that
it can be displayed to users if desired. Some sort of sliding delay
between retries might be a good idea so that the limit isn't burned
through too quickly in certain situations (disconnect ethernet cable,
airplaine mode, etc.).